backend/requirements.txt
Django>=4.2.11
djangorestframework
djongo
djangorestframework-simplejwt
pymongo

